Renames node_config.rs to config.rs
